move to 1 root module, 2 child modules
It didn't make much sense to have the FE and BE be completely independent, so they are deployed together. FE only depends on 1 output from BE, so making changes to either should not really impact the other.

Validated that it deploys, however, everything deployed to us-east-1 which was not intended.

Reduced workflows to 2: build and destroy. Both are currently manual.