Split dev and prod workflows up
because they will do different things in the long term.

Dev will: deploy, test, create a PR
- The pr must be manually accepted
Prod will: deploy, test, delete Dev environment

Removed HostedZone and SiteName from tfvars file because I want to be able to dynamically pass those in based on the environment.

Also added a random int that will get appended to the bucket name so the buckets in prod and dev are the same