The webacl association resource...
is prone to not deploying due to a race condition. By default it will only try to deploy for 5 minutes. The Terraform AWS Provider has added functionality to the resource type to make the creation timeout configurable to address this issue. (see: https://github.com/hashicorp/terraform-provider-aws/pull/30002)

Adjusted the timeout for webacl-association creation to 20 minutes, and increased the overally Terraform deploy workflow timeouts to 45 minutes