Modified Dev Build workflow to be more resilient
If there is already a PR for Test->Main, then it will delete the PR and create a new one that corresponds to the GHA RunID that created it. (This might be completely unnecessary). If there is not a PR in place it will create one. As a result, don't need the Continue-on-error attribute for the PR creation anymore